refactor: change backend service name to backend.ts
